energy-trading-lstm/tests/test_trading.py
def test_trading_strategy():
    import pytest
    from src.models.trading_strategy import smart_trading_strategy
    import numpy as np

    def mock_model(X):
        # Mock model that returns predictable outputs
        return np.array([[10, 5, 20]] * len(X)), np.array([[15]] * len(X))

    defer_demands = [{'start': 0, 'end': 10, 'energy': 5}]
    df_nn_enhanced = pd.DataFrame({
        'avg_buy_price': [10] * 60,
        'avg_sell_price': [15] * 60,
        'demand': [20] * 60,
        'sunlight': [30] * 60
    })

    storage, profit_over_time, actions, satisfied, pending = smart_trading_strategy(
        mock_model, df_nn_enhanced, defer_demands
    )

    assert len(storage) == 60
    assert len(profit_over_time) == 60
    assert 'buy' in actions or 'sell' in actions
    assert len(satisfied) == 1
    assert len(pending) == 0

    # Additional assertions can be added to validate specific logic in trading strategy