energy-trading-lstm/requirements.txt

Flask==2.0.1
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
torch==1.9.0
matplotlib==3.4.2
requests==2.25.1
PyYAML==5.4.1